refactor auditlog into focused subpackages with compatibility aliases
Split storage, store, identity, and status concerns into dedicated sdk/auditlog subpackages and keep root-level API wrappers/aliases so existing call sites continue to compile while improving readability.
