+++ sdk/auditlog/storage/extension.go
@@ -1,38 +1,34 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package log
+package storage
 
 import (
 	"context"
 	"fmt"
 )
 
-type StorageExtension interface {
+type Extension interface {
 	GetClient(ctx context.Context, name string) (StorageClient, error)
 	Start(ctx context.Context) error
 	Shutdown(ctx context.Context) error
 }
 
-type StorageExtensionConfig interface {
+type ExtensionConfig interface {
 	Validate() error
 }
 
-type MemoryStorageExtension struct {
+type MemoryExtension struct {
 	clients map[string]*SimpleKeyValueStorageClient
 }
 
-func NewMemoryStorageExtension() *MemoryStorageExtension {
-	return &MemoryStorageExtension{
-		clients: make(map[string]*SimpleKeyValueStorageClient),
-	}
+func NewMemoryExtension() *MemoryExtension {
+	return &MemoryExtension{clients: make(map[string]*SimpleKeyValueStorageClient)}
 }
 
-func (m *MemoryStorageExtension) Start(ctx context.Context) error {
-	return nil
-}
+func (m *MemoryExtension) Start(ctx context.Context) error { return nil }
 
-func (m *MemoryStorageExtension) Shutdown(ctx context.Context) error {
+func (m *MemoryExtension) Shutdown(ctx context.Context) error {
 	for _, client := range m.clients {
 		if err := client.Close(ctx); err != nil {
 			return err
@@ -41,48 +37,41 @@ func (m *MemoryStorageExtension) Shutdown(ctx context.Context) error {
 	return nil
 }
 
-func (m *MemoryStorageExtension) GetClient(ctx context.Context, name string) (StorageClient, error) {
+func (m *MemoryExtension) GetClient(ctx context.Context, name string) (StorageClient, error) {
 	if client, exists := m.clients[name]; exists {
 		return client, nil
 	}
-
 	client := NewSimpleKeyValueStorageClient()
 	m.clients[name] = client
 	return client, nil
 }
 
-type FileStorageConfig struct {
+type FileConfig struct {
 	Directory string
 }
 
-func (c *FileStorageConfig) Validate() error {
+func (c *FileConfig) Validate() error {
 	if c.Directory == "" {
 		return fmt.Errorf("directory cannot be empty")
 	}
 	return nil
 }
 
-type FileStorageExtension struct {
-	config  *FileStorageConfig
+type FileExtension struct {
+	config  *FileConfig
 	clients map[string]*BoltDBStorageClient
 }
 
-func NewFileStorageExtension(config *FileStorageConfig) (*FileStorageExtension, error) {
+func NewFileExtension(config *FileConfig) (*FileExtension, error) {
 	if err := config.Validate(); err != nil {
 		return nil, err
 	}
-
-	return &FileStorageExtension{
-		config:  config,
-		clients: make(map[string]*BoltDBStorageClient),
-	}, nil
+	return &FileExtension{config: config, clients: make(map[string]*BoltDBStorageClient)}, nil
 }
 
-func (f *FileStorageExtension) Start(ctx context.Context) error {
-	return nil
-}
+func (f *FileExtension) Start(ctx context.Context) error { return nil }
 
-func (f *FileStorageExtension) Shutdown(ctx context.Context) error {
+func (f *FileExtension) Shutdown(ctx context.Context) error {
 	for _, client := range f.clients {
 		if err := client.Close(ctx); err != nil {
 			return err
@@ -91,42 +80,34 @@ func (f *FileStorageExtension) Shutdown(ctx context.Context) error {
 	return nil
 }
 
-func (f *FileStorageExtension) GetClient(ctx context.Context, name string) (StorageClient, error) {
+func (f *FileExtension) GetClient(ctx context.Context, name string) (StorageClient, error) {
 	if client, exists := f.clients[name]; exists {
 		return client, nil
 	}
-
 	filePath := fmt.Sprintf("%s/%s.db", f.config.Directory, name)
 	client, err := NewBoltDBStorageClient(filePath)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create file storage client: %w", err)
 	}
-
 	f.clients[name] = client
 	return client, nil
 }
 
-type RedisStorageExtension struct {
+type RedisExtension struct {
 	config  RedisStorageConfig
 	clients map[string]*RealRedisStorageClient
 }
 
-func NewRedisStorageExtension(config RedisStorageConfig) (*RedisStorageExtension, error) {
+func NewRedisExtension(config RedisStorageConfig) (*RedisExtension, error) {
 	if config.Endpoint == "" {
 		return nil, fmt.Errorf("endpoint cannot be empty")
 	}
-
-	return &RedisStorageExtension{
-		config:  config,
-		clients: make(map[string]*RealRedisStorageClient),
-	}, nil
+	return &RedisExtension{config: config, clients: make(map[string]*RealRedisStorageClient)}, nil
 }
 
-func (r *RedisStorageExtension) Start(ctx context.Context) error {
-	return nil
-}
+func (r *RedisExtension) Start(ctx context.Context) error { return nil }
 
-func (r *RedisStorageExtension) Shutdown(ctx context.Context) error {
+func (r *RedisExtension) Shutdown(ctx context.Context) error {
 	for _, client := range r.clients {
 		if err := client.Close(ctx); err != nil {
 			return err
@@ -135,47 +116,38 @@ func (r *RedisStorageExtension) Shutdown(ctx context.Context) error {
 	return nil
 }
 
-func (r *RedisStorageExtension) GetClient(ctx context.Context, name string) (StorageClient, error) {
+func (r *RedisExtension) GetClient(ctx context.Context, name string) (StorageClient, error) {
 	if client, exists := r.clients[name]; exists {
 		return client, nil
 	}
-
 	clientConfig := r.config
 	clientConfig.Prefix = r.config.Prefix + name + "_"
-
 	client, err := NewRealRedisStorageClient(clientConfig)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create Redis storage client: %w", err)
 	}
-
 	r.clients[name] = client
 	return client, nil
 }
 
-type SQLStorageExtension struct {
+type SQLExtension struct {
 	config  SQLStorageConfig
 	clients map[string]*SQLStorageClient
 }
 
-func NewSQLStorageExtension(config SQLStorageConfig) (*SQLStorageExtension, error) {
+func NewSQLExtension(config SQLStorageConfig) (*SQLExtension, error) {
 	if config.Driver == "" {
 		return nil, fmt.Errorf("driver cannot be empty")
 	}
 	if config.Datasource == "" {
 		return nil, fmt.Errorf("datasource cannot be empty")
 	}
-
-	return &SQLStorageExtension{
-		config:  config,
-		clients: make(map[string]*SQLStorageClient),
-	}, nil
+	return &SQLExtension{config: config, clients: make(map[string]*SQLStorageClient)}, nil
 }
 
-func (s *SQLStorageExtension) Start(ctx context.Context) error {
-	return nil
-}
+func (s *SQLExtension) Start(ctx context.Context) error { return nil }
 
-func (s *SQLStorageExtension) Shutdown(ctx context.Context) error {
+func (s *SQLExtension) Shutdown(ctx context.Context) error {
 	for _, client := range s.clients {
 		if err := client.Close(ctx); err != nil {
 			return err
@@ -184,19 +156,16 @@ func (s *SQLStorageExtension) Shutdown(ctx context.Context) error {
 	return nil
 }
 
-func (s *SQLStorageExtension) GetClient(ctx context.Context, name string) (StorageClient, error) {
+func (s *SQLExtension) GetClient(ctx context.Context, name string) (StorageClient, error) {
 	if client, exists := s.clients[name]; exists {
 		return client, nil
 	}
-
 	clientConfig := s.config
 	clientConfig.TableName = s.config.TableName + "_" + name
-
 	client, err := NewSQLStorageClient(clientConfig)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create SQL storage client: %w", err)
 	}
-
 	s.clients[name] = client
 	return client, nil
 }
