+++ sdk/auditlog/storage/redis_client.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package log
+package storage
 
 import (
 	"context"
@@ -21,30 +21,21 @@ func NewRealRedisStorageClient(config RedisStorageConfig) (*RealRedisStorageClie
 	if config.Endpoint == "" {
 		return nil, fmt.Errorf("endpoint cannot be empty")
 	}
-
 	client := redis.NewClient(&redis.Options{
 		Addr:     config.Endpoint,
 		Password: config.Password,
 		DB:       config.DB,
 	})
-
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
-
 	if err := client.Ping(ctx).Err(); err != nil {
 		return nil, fmt.Errorf("failed to connect to Redis at %s: %w", config.Endpoint, err)
 	}
-
-	return &RealRedisStorageClient{
-		client:     client,
-		prefix:     config.Prefix,
-		expiration: config.Expiration,
-	}, nil
+	return &RealRedisStorageClient{client: client, prefix: config.Prefix, expiration: config.Expiration}, nil
 }
 
 func (c *RealRedisStorageClient) Get(ctx context.Context, key string) ([]byte, error) {
 	fullKey := c.prefix + key
-	
 	result, err := c.client.Get(ctx, fullKey).Bytes()
 	if err == redis.Nil {
 		return nil, fmt.Errorf("key not found: %s", key)
@@ -52,38 +43,31 @@ func (c *RealRedisStorageClient) Get(ctx context.Context, key string) ([]byte, e
 	if err != nil {
 		return nil, fmt.Errorf("failed to get key %s: %w", key, err)
 	}
-
 	return result, nil
 }
 
 func (c *RealRedisStorageClient) Set(ctx context.Context, key string, value []byte) error {
 	fullKey := c.prefix + key
-
 	expiration := c.expiration
 	if expiration == 0 {
 		expiration = 24 * time.Hour
 	}
-
 	if err := c.client.Set(ctx, fullKey, value, expiration).Err(); err != nil {
 		return fmt.Errorf("failed to set key %s: %w", key, err)
 	}
-
 	return nil
 }
 
 func (c *RealRedisStorageClient) Delete(ctx context.Context, key string) error {
 	fullKey := c.prefix + key
-
 	if err := c.client.Del(ctx, fullKey).Err(); err != nil {
 		return fmt.Errorf("failed to delete key %s: %w", key, err)
 	}
-
 	return nil
 }
 
 func (c *RealRedisStorageClient) Batch(ctx context.Context, ops ...Operation) error {
 	pipe := c.client.Pipeline()
-
 	for _, op := range ops {
 		switch o := op.(type) {
 		case *SetOperation:
@@ -98,11 +82,9 @@ func (c *RealRedisStorageClient) Batch(ctx context.Context, ops ...Operation) er
 			pipe.Del(ctx, fullKey)
 		}
 	}
-
 	if _, err := pipe.Exec(ctx); err != nil {
 		return fmt.Errorf("failed to execute batch operations: %w", err)
 	}
-
 	return nil
 }
 
@@ -112,4 +94,3 @@ func (c *RealRedisStorageClient) Close(ctx context.Context) error {
 	}
 	return nil
 }
-
