+++ sdk/auditlog/storage/types.go
@@ -1,0 +1,278 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package storage
+
+import (
+	"context"
+	"fmt"
+	"sync"
+	"time"
+)
+
+type StorageClient interface {
+	Get(ctx context.Context, key string) ([]byte, error)
+	Set(ctx context.Context, key string, value []byte) error
+	Delete(ctx context.Context, key string) error
+	Batch(ctx context.Context, ops ...Operation) error
+	Close(ctx context.Context) error
+}
+
+type Operation interface {
+	Execute(ctx context.Context) error
+}
+
+type SetOperation struct {
+	Key   string
+	Value []byte
+}
+
+func (op *SetOperation) Execute(ctx context.Context) error {
+	return nil
+}
+
+type DeleteOperation struct {
+	Key string
+}
+
+func (op *DeleteOperation) Execute(ctx context.Context) error {
+	return nil
+}
+
+type SimpleKeyValueStorageClient struct {
+	storage map[string][]byte
+	mutex   sync.RWMutex
+}
+
+func NewSimpleKeyValueStorageClient() *SimpleKeyValueStorageClient {
+	return &SimpleKeyValueStorageClient{storage: make(map[string][]byte)}
+}
+
+func (c *SimpleKeyValueStorageClient) Get(ctx context.Context, key string) ([]byte, error) {
+	c.mutex.RLock()
+	defer c.mutex.RUnlock()
+	value, exists := c.storage[key]
+	if !exists {
+		return nil, fmt.Errorf("key not found: %s", key)
+	}
+	result := make([]byte, len(value))
+	copy(result, value)
+	return result, nil
+}
+
+func (c *SimpleKeyValueStorageClient) Set(ctx context.Context, key string, value []byte) error {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	valueCopy := make([]byte, len(value))
+	copy(valueCopy, value)
+	c.storage[key] = valueCopy
+	return nil
+}
+
+func (c *SimpleKeyValueStorageClient) Delete(ctx context.Context, key string) error {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	delete(c.storage, key)
+	return nil
+}
+
+func (c *SimpleKeyValueStorageClient) Batch(ctx context.Context, ops ...Operation) error {
+	for _, op := range ops {
+		if err := op.Execute(ctx); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (c *SimpleKeyValueStorageClient) Close(ctx context.Context) error {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	c.storage = make(map[string][]byte)
+	return nil
+}
+
+func (c *SimpleKeyValueStorageClient) Clear() {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	c.storage = make(map[string][]byte)
+}
+
+func (c *SimpleKeyValueStorageClient) Size() int {
+	c.mutex.RLock()
+	defer c.mutex.RUnlock()
+	return len(c.storage)
+}
+
+type BoltDBStorageClient struct {
+	filePath string
+	mutex    sync.RWMutex
+	storage  map[string][]byte
+}
+
+func NewBoltDBStorageClient(filePath string) (*BoltDBStorageClient, error) {
+	if filePath == "" {
+		return nil, fmt.Errorf("file path cannot be empty")
+	}
+	client := &BoltDBStorageClient{
+		filePath: filePath,
+		storage:  make(map[string][]byte),
+	}
+	if err := client.load(context.Background()); err != nil {
+		return nil, fmt.Errorf("failed to load storage: %w", err)
+	}
+	return client, nil
+}
+
+func (c *BoltDBStorageClient) Get(ctx context.Context, key string) ([]byte, error) {
+	c.mutex.RLock()
+	defer c.mutex.RUnlock()
+	value, exists := c.storage[key]
+	if !exists {
+		return nil, fmt.Errorf("key not found: %s", key)
+	}
+	result := make([]byte, len(value))
+	copy(result, value)
+	return result, nil
+}
+
+func (c *BoltDBStorageClient) Set(ctx context.Context, key string, value []byte) error {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	valueCopy := make([]byte, len(value))
+	copy(valueCopy, value)
+	c.storage[key] = valueCopy
+	return c.persist(ctx)
+}
+
+func (c *BoltDBStorageClient) Delete(ctx context.Context, key string) error {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	delete(c.storage, key)
+	return c.persist(ctx)
+}
+
+func (c *BoltDBStorageClient) Batch(ctx context.Context, ops ...Operation) error {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	for _, op := range ops {
+		switch o := op.(type) {
+		case *SetOperation:
+			valueCopy := make([]byte, len(o.Value))
+			copy(valueCopy, o.Value)
+			c.storage[o.Key] = valueCopy
+		case *DeleteOperation:
+			delete(c.storage, o.Key)
+		}
+	}
+	return c.persist(ctx)
+}
+
+func (c *BoltDBStorageClient) Close(ctx context.Context) error {
+	return c.persist(ctx)
+}
+
+func (c *BoltDBStorageClient) load(ctx context.Context) error { return nil }
+func (c *BoltDBStorageClient) persist(ctx context.Context) error { return nil }
+
+type RedisStorageConfig struct {
+	Endpoint   string
+	Password   string
+	DB         int
+	Prefix     string
+	Expiration time.Duration
+}
+
+func NewRedisStorageClient(config RedisStorageConfig) (StorageClient, error) {
+	return NewRealRedisStorageClient(config)
+}
+
+type SQLStorageConfig struct {
+	Driver     string
+	Datasource string
+	TableName  string
+}
+
+type SQLStorageClient struct {
+	driver     string
+	datasource string
+	tableName  string
+	mutex      sync.RWMutex
+	storage    map[string][]byte
+}
+
+func NewSQLStorageClient(config SQLStorageConfig) (*SQLStorageClient, error) {
+	if config.Driver == "" {
+		return nil, fmt.Errorf("driver cannot be empty")
+	}
+	if config.Datasource == "" {
+		return nil, fmt.Errorf("datasource cannot be empty")
+	}
+	tableName := config.TableName
+	if tableName == "" {
+		tableName = "audit_logs"
+	}
+	client := &SQLStorageClient{
+		driver:     config.Driver,
+		datasource: config.Datasource,
+		tableName:  tableName,
+		storage:    make(map[string][]byte),
+	}
+	if err := client.initialize(context.Background()); err != nil {
+		return nil, fmt.Errorf("failed to initialize SQL storage: %w", err)
+	}
+	return client, nil
+}
+
+func (c *SQLStorageClient) initialize(ctx context.Context) error { return nil }
+
+func (c *SQLStorageClient) Get(ctx context.Context, key string) ([]byte, error) {
+	c.mutex.RLock()
+	defer c.mutex.RUnlock()
+	value, exists := c.storage[key]
+	if !exists {
+		return nil, fmt.Errorf("key not found: %s", key)
+	}
+	result := make([]byte, len(value))
+	copy(result, value)
+	return result, nil
+}
+
+func (c *SQLStorageClient) Set(ctx context.Context, key string, value []byte) error {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	valueCopy := make([]byte, len(value))
+	copy(valueCopy, value)
+	c.storage[key] = valueCopy
+	return nil
+}
+
+func (c *SQLStorageClient) Delete(ctx context.Context, key string) error {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	delete(c.storage, key)
+	return nil
+}
+
+func (c *SQLStorageClient) Batch(ctx context.Context, ops ...Operation) error {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	for _, op := range ops {
+		switch o := op.(type) {
+		case *SetOperation:
+			valueCopy := make([]byte, len(o.Value))
+			copy(valueCopy, o.Value)
+			c.storage[o.Key] = valueCopy
+		case *DeleteOperation:
+			delete(c.storage, o.Key)
+		}
+	}
+	return nil
+}
+
+func (c *SQLStorageClient) Close(ctx context.Context) error {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	c.storage = make(map[string][]byte)
+	return nil
+}
