+++ sdk/auditlog/store/store.go
@@ -1,0 +1,311 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package store
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+	"sync"
+
+	"go.opentelemetry.io/otel/sdk/auditlog/identity"
+	"go.opentelemetry.io/otel/sdk/auditlog/status"
+	sdklog "go.opentelemetry.io/otel/sdk/log"
+)
+
+type AuditLogStore interface {
+	Save(ctx context.Context, record *sdklog.Record) error
+	RemoveAll(ctx context.Context, records []sdklog.Record) error
+	GetAll(ctx context.Context) ([]sdklog.Record, error)
+}
+
+type AuditLogFileStore struct {
+	logFilePath string
+	loggedRecords map[string]string
+	mutex sync.RWMutex
+	defaultLogFileName string
+	logFileExtension string
+}
+
+const (
+	DefaultLogFileExtension = ".log"
+	DefaultLogFileName = "audit" + DefaultLogFileExtension
+)
+
+func NewAuditLogFileStore(path string) (*AuditLogFileStore, error) {
+	store := &AuditLogFileStore{
+		loggedRecords: make(map[string]string),
+		defaultLogFileName: DefaultLogFileName,
+		logFileExtension: DefaultLogFileExtension,
+	}
+	if info, err := os.Stat(path); err == nil && info.IsDir() {
+		store.logFilePath = filepath.Join(path, DefaultLogFileName)
+	} else {
+		store.logFilePath = path
+	}
+	if err := os.MkdirAll(filepath.Dir(store.logFilePath), 0755); err != nil {
+		return nil, fmt.Errorf("failed to create parent directories: %w", err)
+	}
+	if _, err := os.Stat(store.logFilePath); os.IsNotExist(err) {
+		file, err := os.Create(store.logFilePath)
+		if err != nil {
+			return nil, fmt.Errorf("failed to create log file: %w", err)
+		}
+		file.Close()
+	}
+	if err := store.verifyFileAccess(); err != nil {
+		return nil, err
+	}
+	if err := store.loadExistingRecordIds(context.Background()); err != nil {
+		return nil, fmt.Errorf("failed to load existing record IDs: %w", err)
+	}
+	return store, nil
+}
+
+func (s *AuditLogFileStore) verifyFileAccess() error {
+	file, err := os.Open(s.logFilePath)
+	if err != nil {
+		return fmt.Errorf("cannot read %s: %w", s.logFilePath, err)
+	}
+	file.Close()
+	file, err = os.OpenFile(s.logFilePath, os.O_WRONLY|os.O_APPEND, 0644)
+	if err != nil {
+		return fmt.Errorf("cannot write to %s: %w", s.logFilePath, err)
+	}
+	file.Close()
+	return nil
+}
+
+func (s *AuditLogFileStore) LogFilePath() string {
+	return s.logFilePath
+}
+
+func (s *AuditLogFileStore) loadExistingRecordIds(ctx context.Context) error {
+	file, err := os.Open(s.logFilePath)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+	if info, err := file.Stat(); err != nil || info.Size() == 0 {
+		return nil
+	}
+	decoder := json.NewDecoder(file)
+	for {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		default:
+		}
+		var record sdklog.Record
+		if err := decoder.Decode(&record); err != nil {
+			if err == io.EOF {
+				break
+			}
+			continue
+		}
+		recordID, err := identity.GetRecordID(&record)
+		if err == nil {
+			s.loggedRecords[recordID] = identity.GetRecordHash(&record)
+		}
+	}
+	return nil
+}
+
+func (s *AuditLogFileStore) Save(ctx context.Context, record *sdklog.Record) error {
+	if record == nil {
+		return fmt.Errorf("record cannot be nil")
+	}
+	recordID, err := identity.GetRecordID(record)
+	if err != nil {
+		return err
+	}
+	s.mutex.Lock()
+	defer s.mutex.Unlock()
+	recordHash := identity.GetRecordHash(record)
+	if existingHash, exists := s.loggedRecords[recordID]; exists {
+		if existingHash != "" && recordHash != "" && existingHash != recordHash {
+			return status.New(status.ErrorConflict, "duplicate record_id with different hash", false, nil)
+		}
+		return nil
+	}
+	file, err := os.OpenFile(s.logFilePath, os.O_WRONLY|os.O_APPEND, 0644)
+	if err != nil {
+		return fmt.Errorf("failed to open log file for writing: %w", err)
+	}
+	defer file.Close()
+	encoder := json.NewEncoder(file)
+	if err := encoder.Encode(record); err != nil {
+		return fmt.Errorf("failed to encode record: %w", err)
+	}
+	if _, err := file.WriteString("\n"); err != nil {
+		return fmt.Errorf("failed to write newline: %w", err)
+	}
+	s.loggedRecords[recordID] = recordHash
+	return nil
+}
+
+func (s *AuditLogFileStore) RemoveAll(ctx context.Context, records []sdklog.Record) error {
+	if len(records) == 0 {
+		return nil
+	}
+	s.mutex.Lock()
+	defer s.mutex.Unlock()
+	recordIDsToRemove := make(map[string]bool)
+	for _, record := range records {
+		recordID, err := identity.GetRecordID(&record)
+		if err == nil {
+			recordIDsToRemove[recordID] = true
+		}
+	}
+	file, err := os.Open(s.logFilePath)
+	if err != nil {
+		return fmt.Errorf("failed to open log file for reading: %w", err)
+	}
+	defer file.Close()
+	var remainingRecords []sdklog.Record
+	decoder := json.NewDecoder(file)
+	for {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		default:
+		}
+		var record sdklog.Record
+		if err := decoder.Decode(&record); err != nil {
+			if err == io.EOF {
+				break
+			}
+			continue
+		}
+		recordID, err := identity.GetRecordID(&record)
+		if err != nil {
+			continue
+		}
+		if !recordIDsToRemove[recordID] {
+			remainingRecords = append(remainingRecords, record)
+		} else {
+			delete(s.loggedRecords, recordID)
+		}
+	}
+	tempFile, err := os.CreateTemp(filepath.Dir(s.logFilePath), "audit_temp_*")
+	if err != nil {
+		return fmt.Errorf("failed to create temp file: %w", err)
+	}
+	defer os.Remove(tempFile.Name())
+	encoder := json.NewEncoder(tempFile)
+	for _, record := range remainingRecords {
+		if err := encoder.Encode(record); err != nil {
+			tempFile.Close()
+			return fmt.Errorf("failed to encode record: %w", err)
+		}
+		if _, err := tempFile.WriteString("\n"); err != nil {
+			tempFile.Close()
+			return fmt.Errorf("failed to write newline: %w", err)
+		}
+	}
+	if err := tempFile.Close(); err != nil {
+		return fmt.Errorf("failed to close temp file: %w", err)
+	}
+	if err := os.Rename(tempFile.Name(), s.logFilePath); err != nil {
+		return fmt.Errorf("failed to replace original file: %w", err)
+	}
+	return nil
+}
+
+func (s *AuditLogFileStore) GetAll(ctx context.Context) ([]sdklog.Record, error) {
+	s.mutex.RLock()
+	defer s.mutex.RUnlock()
+	file, err := os.Open(s.logFilePath)
+	if err != nil {
+		return nil, fmt.Errorf("failed to open log file: %w", err)
+	}
+	defer file.Close()
+	var records []sdklog.Record
+	decoder := json.NewDecoder(file)
+	for {
+		select {
+		case <-ctx.Done():
+			return nil, ctx.Err()
+		default:
+		}
+		var record sdklog.Record
+		if err := decoder.Decode(&record); err != nil {
+			if err == io.EOF {
+				break
+			}
+			continue
+		}
+		records = append(records, record)
+	}
+	return records, nil
+}
+
+type AuditLogInMemoryStore struct {
+	records []sdklog.Record
+	mutex   sync.RWMutex
+}
+
+func NewAuditLogInMemoryStore() *AuditLogInMemoryStore {
+	return &AuditLogInMemoryStore{records: make([]sdklog.Record, 0)}
+}
+
+func (s *AuditLogInMemoryStore) Save(ctx context.Context, record *sdklog.Record) error {
+	if record == nil {
+		return fmt.Errorf("record cannot be nil")
+	}
+	s.mutex.Lock()
+	defer s.mutex.Unlock()
+	recordCopy := record.Clone()
+	s.records = append(s.records, recordCopy)
+	return nil
+}
+
+func (s *AuditLogInMemoryStore) RemoveAll(ctx context.Context, records []sdklog.Record) error {
+	if len(records) == 0 {
+		return nil
+	}
+	s.mutex.Lock()
+	defer s.mutex.Unlock()
+	recordsToRemove := make(map[string]bool)
+	for _, record := range records {
+		recordCopy := record
+		id, err := identity.GetRecordID(&recordCopy)
+		if err == nil {
+			recordsToRemove[id] = true
+		}
+	}
+	var filteredRecords []sdklog.Record
+	for _, record := range s.records {
+		recordCopy := record
+		id, err := identity.GetRecordID(&recordCopy)
+		if err != nil || !recordsToRemove[id] {
+			filteredRecords = append(filteredRecords, record)
+		}
+	}
+	s.records = filteredRecords
+	return nil
+}
+
+func (s *AuditLogInMemoryStore) GetAll(ctx context.Context) ([]sdklog.Record, error) {
+	s.mutex.RLock()
+	defer s.mutex.RUnlock()
+	records := make([]sdklog.Record, len(s.records))
+	copy(records, s.records)
+	return records, nil
+}
+
+func (s *AuditLogInMemoryStore) GetRecordCount() int {
+	s.mutex.RLock()
+	defer s.mutex.RUnlock()
+	return len(s.records)
+}
+
+func (s *AuditLogInMemoryStore) Clear() {
+	s.mutex.Lock()
+	defer s.mutex.Unlock()
+	s.records = make([]sdklog.Record, 0)
+}
